iccu/requirements.txt
polars[pyarrow,excel]>=1.12.0,<2
pandas[xml]>=2.2.3,<3
osm_conflate==1.4.1
phonenumberslite>=8.13.49,<9
